moved activites that are not main or mapview to subactivities/ to clean up the top-level of the project.
