+++ android/src/main/java/org/mozilla/mozstumbler/client/subactivities/LogActivity.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.mozstumbler.client;
+package org.mozilla.mozstumbler.client.subactivities;
 
 import android.app.Activity;
 import android.content.BroadcastReceiver;
