+++ android/src/main/res/layout/activity_leaderboard.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="org.mozilla.mozstumbler.client.LeaderboardActivity">
+    tools:context="org.mozilla.mozstumbler.client.subactivities.LeaderboardActivity">
 
     <WebView
         android:id="@+id/webview"
