+++ android/src/main/res/layout/activity_log.xml
@@ -8,7 +8,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context=".LogActivity">
     <view
-        class="org.mozilla.mozstumbler.client.LogActivity$ConsoleView"
+        class="org.mozilla.mozstumbler.client.subactivities.LogActivity$ConsoleView"
         android:id="@+id/scrollview"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent" />
